+++ vnext/IntegrationTests/TestInstance.h
@@ -86,8 +86,8 @@ class TestNativeUIManager : public facebook::react::INativeUIManager {
       facebook::xplat::module::CxxModule::Callback callback) override;
   void findSubviewIn(
       facebook::react::ShadowNode &shadowNode,
-      int64_t x,
-      int64_t y,
+      float x,
+      float y,
       facebook::xplat::module::CxxModule::Callback callback) override;
   void focus(int64_t reactTag) override;
   void blur(int64_t reactTag) override;
