+++ vnext/ReactUWP/Modules/NativeUIManager.h
@@ -72,8 +72,8 @@ class NativeUIManager : public facebook::react::INativeUIManager {
       facebook::xplat::module::CxxModule::Callback callback) override;
   void findSubviewIn(
       facebook::react::ShadowNode &shadowNode,
-      int64_t x,
-      int64_t y,
+      float x,
+      float y,
       facebook::xplat::module::CxxModule::Callback callback) override;
 
   void focus(int64_t reactTag) override;
